+++ src/pages/bee_farm/MyFarmsPage.js
@@ -1,8 +1,8 @@
 import React from "react";
-import MainMenu from "../components/menu/MainMenu"
+import MainMenu from "../../components/menu/MainMenu"
 import {Button, Card, Container, Grid, Icon, Modal, Segment} from "semantic-ui-react";
-import CreateBeeFarmForm from "../components/forms/bee_farm/create/CreateBeeFarmForm";
-import {GET_API} from "../http/GET_API";
+import CreateBeeFarmForm from "../../components/forms/bee_farm/create/CreateBeeFarmForm";
+import {GET_API} from "../../http/GET_API";
 
 /*
  Страница выбора пользователем пасеки.
