+++ src/pages/harvests/ControlHarvestPage.js
@@ -1,19 +1,19 @@
 import React from "react";
-import MainMenu from "../components/menu/MainMenu"
+import MainMenu from "../../components/menu/MainMenu"
 import {Button, Container, Grid, Icon, Loader, Modal, Pagination, Segment, Table} from "semantic-ui-react";
-import CreatePollenHarvestForm from "../components/forms/harvests/CreatePollenHarvestForm";
-import DeleteModal from "../components/modal/DeleteModal";
-import TablePaginatorComponent from "../components/other/TablePaginatorComponent";
+import CreateControlHarvestForm from "../../components/forms/harvests/CreateControlHarvestForm";
+import DeleteModal from "../../components/modal/DeleteModal";
+import TablePaginatorComponent from "../../components/other/TablePaginatorComponent";
 
 /*
- Страница сборов пыльцы
+ Страница контрольных сборов
  */
-class PollenHarvestPage extends TablePaginatorComponent {
+class ControlHarvestPage extends TablePaginatorComponent {
     constructor(props) {
         super(props);
 
-        this.getItemsCallback = this.getAPI.GetUsersPollenHarvests;
-        this.deleteItemCallback = this.deleteAPI.DeletePollenHarvestByID;
+        this.getItemsCallback = this.getAPI.GetUsersControlHarvests;
+        this.deleteItemCallback = this.deleteAPI.DeleteControlHarvestByID;
     }
 
     render() {
@@ -24,7 +24,7 @@ class PollenHarvestPage extends TablePaginatorComponent {
                     <Grid>
                         <Grid.Row columns={2} relaxed='very'>
                             <Grid.Column>
-                                <h1 style={{textAlign: "center"}}>Сборы пыльцы</h1>
+                                <h1 style={{textAlign: "center"}}>Контрольные сборы мёда</h1>
                             </Grid.Column>
                             <Grid.Column>
                                 <Modal trigger={<Button
@@ -36,9 +36,9 @@ class PollenHarvestPage extends TablePaginatorComponent {
                                     label={{ basic: true, color: 'green', pointing: 'left', content: this.state.itemsCount }}
                                     style={{marginRight: "30px"}}
                                 />}>
-                                    <Modal.Header>Добавить сбор</Modal.Header>
+                                    <Modal.Header>Новый сбор</Modal.Header>
                                     <Modal.Content>
-                                        <CreatePollenHarvestForm />
+                                        <CreateControlHarvestForm />
                                     </Modal.Content>
                                 </Modal>
                             </Grid.Column>
@@ -47,8 +47,8 @@ class PollenHarvestPage extends TablePaginatorComponent {
                     <Table celled>
                         <Table.Header>
                             <Table.Row>
-                                <Table.HeaderCell>Пасека&nbsp;
-                                    <Icon link name='arrow down' onClick={this.reorder.bind(this, "bee_farm_id")} />
+                                <Table.HeaderCell>Семья&nbsp;
+                                    <Icon link name='arrow down' onClick={this.reorder.bind(this, "bee_family_id")} />
                                 </Table.HeaderCell>
                                 <Table.HeaderCell>Дата&nbsp;
                                     <Icon link name='arrow down' onClick={this.reorder.bind(this, "date")} />
@@ -63,10 +63,10 @@ class PollenHarvestPage extends TablePaginatorComponent {
                         <Table.Body>
                             {this.state.items === null ? <Table.Row>
                                     <Table.Cell><Loader active inline /></Table.Cell>
-                            </Table.Row> :
+                                </Table.Row> :
                                 this.state.items.map((item) => {
                                     return <Table.Row key={item.id}>
-                                        <Table.Cell>{item["bee_farm"].name}</Table.Cell>
+                                        <Table.Cell>{item["bee_family"].name}</Table.Cell>
                                         <Table.Cell>
                                             {item["date"] === null ? '' :
                                                 (new Date(item["date"])).toLocaleString('ru', {
@@ -96,4 +96,4 @@ class PollenHarvestPage extends TablePaginatorComponent {
     }
 }
 
-export default PollenHarvestPage;
+export default ControlHarvestPage;
