+++ src/pages/wiki/WikiIndexPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import MainMenu from "../components/menu/MainMenu"
+import MainMenu from "../../components/menu/MainMenu"
 import {Container, Message, Grid} from "semantic-ui-react";
 
 
@@ -14,7 +14,7 @@ class WikiIndexPage extends React.Component {
     render() {
         return <div>
             <Container>
-                <MainMenu activeItem={'Вики'} history={this.props.history} />
+                <MainMenu activeItem={'Справочник'} history={this.props.history} />
                 <Grid centered columns={4}>
                     <Grid.Column>
                         <Message compact>
